Remove file opening from funLib
