+++ userinc/funLib.h
@@ -58,8 +58,7 @@ void printCuts(cutsValues &cutsDefinition);
 
 std::map<std::string,std::string> parseOptions(std::string s);
 const std::vector<std::string> tokenize(std::string s, const char delim);
-bool isFilteredEvent(int nrun, int nburst, int timestamp, vector<eventID> &badEventsList);
-int common_init(std::string filePrefix, std::string filterFile, vector<eventID> &badEventsList);
+bool isFilteredEvent(int nrun, int nburst, int timestamp, const vector<eventID> &badEventsList);
 
 extern double Mpi0;
 extern double Mpic;
